be/requirements.txt
alembic==1.9.2
async-timeout==4.0.3
attrs==22.2.0
blinker==1.9.0
Brotli==1.0.9
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cssselect2==0.7.0
dnspython==2.3.0
email_validator==2.2.0
Flask==3.1.0
Flask-Cors==5.0.0
Flask-JWT-Extended==4.6.0
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.0.5
Flask-Principal==0.4.0
Flask-Security-Too==5.3.3
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.2.2
fonttools==4.40.0
greenlet==2.0.2
gunicorn==23.0.0
html5lib==1.1
idna==3.4
importlib-resources==6.1.1
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.2
loguru==0.7.2
lxml==4.9.2
Mako==1.2.4
MarkupSafe==2.1.2
marshmallow==3.19.0
mistune==3.0.2
numpy==2.1.1
packaging==23.0
passlib==1.7.4
pillow==11.0.0
platformdirs==3.1.1
psycopg2-binary==2.9.5
pycparser==2.21
PyJWT==2.6.0
PyMySQL==1.0.2
pyodbc==4.0.39
pyphen==0.14.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
pytz==2022.7.1
PyYAML==6.0.2
referencing==0.35.1
requests==2.28.2
requests-file==1.5.1
requests-toolbelt==1.0.0
rpds-py==0.21.0
six==1.16.0
SQLAlchemy==2.0.5.post1
SQLAlchemy-Utils==0.41.2
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
webargs==8.2.0
webencodings==0.5.1
Werkzeug==3.1.3
WTForms==3.2.1
zeep==4.2.1
zopfli==0.2.2
